# Convert JsonToken to readonly struct