JsonToken pops quite high in memory allocations.
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/bfbb8cdf-59f5-45a8-a6e6-45070b32afa4)

This PR will convert it to readonly struct as it is already designed as immutable sealed class.