@TrayanZapryanov 2024-02-12
Unfortunately I cannot build it on my machine.
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/4fa946be-a6b7-40e9-a013-985d49335397)