@jskeet 2024-02-12
Eek, you can't build anything that needs nuget.org? Okay, I can pull and look locally - but it won't be terribly soon.