@TrayanZapryanov 2024-02-14
@jskeet Can you start build also here?
So I can see the failing test.