@jskeet 2024-02-14
Trying to fix code via CI (with me having to restart the build each time you make a change) isn't going to scale. Please, see if you can find a way of configuring your machine to build the code you're proposing (or do so in a VM etc).