@TrayanZapryanov 2024-02-15
@jskeet I've found what is preventing me from clean build.
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/8df03a2f-231d-4a20-b081-5ca9f88fec2b)

If this is related to Unix only build maybe it is a good idea to add OS condition like suggested [here](https://stackoverflow.com/questions/52612790/different-nuget-package-based-on-operating-system)