@jskeet 2024-02-15
@TrayanZapryanov That shouldn't stop you building (I'm building on Windows), but you may need to reconfigure package sources.