@TrayanZapryanov 2024-02-15
@jskeet I've managed to build on my machine by commenting this.
Also managed to run and fixt tests.
Please check if this is okay for you.

Also remove nullable annotations from other PR