@jskeet 2024-02-27
I've tried to reproduce benefits with this in benchmarks, and I'm not seeing anything. (I think the StringBuilder caching is a clearer win.) Have you tried running your own tests with *just* this change (without the StringBuilder caching)?