@TrayanZapryanov 2024-03-01
@jskeet Here are .Memory results with this change:
Before:
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/6e3870e2-3738-4ecd-be0b-081992779f57)

After:
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/2e89fb8c-48f1-45ac-94b0-dd5a449ae91b)


We have pretty big "proto" json definitions of workflows stored into database and my test is just deserializing all of them.
As you say - benefit is not quite big - only 20MB, but in general change looks quite straight forward and safe. But it is up to you to decide if it fits your bar.