@TrayanZapryanov 2024-03-01
This are results from .Trace
Before:
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/9b939740-dc5c-4932-ac2c-3c2cba50c873)

After:
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/97ddec44-c227-44a7-81ff-849254156f4e)

There are some improvements, but nothing dramatical.