@TrayanZapryanov 2024-03-06
I've build a benchmark program and also can confirm a small performance drop in this changes.
Here it is a benchmark that I use:

```
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using Google.Protobuf;

BenchmarkSwitcher.FromAssembly(typeof(Tests).Assembly).Run(args);

[MemoryDiagnoser(false)]
[HideColumns("Job", "Error", "StdDev", "Median", "RatioSD")]
public class Tests
{
	private string _smallJson;
	private string _bigJson;

	[GlobalSetup]
	public void Setup()
	{
		_smallJson = "{ \"foo\": { \"x\": 1, \"y\": [ 0 ] } }";
		_bigJson = File.ReadAllText("BigDefinition.json");
	}

	[Benchmark]
	public int Small()
	{
		var tokenizer = JsonTokenizer.FromTextReader(new StringReader(_smallJson));
		var tokens = 0;
		while (tokenizer.Next() != JsonToken.EndDocument)
		{
			tokens++;
		}

		return tokens;
	}

	[Benchmark]
	public int Big()
	{
		var tokenizer = JsonTokenizer.FromTextReader(new StringReader(_bigJson));
		var tokens = 0;
		while (tokenizer.Next() != JsonToken.EndDocument)
		{
			tokens++;
		}

		return tokens;
	}
}
```

one of our test definitions that we are deserializing is the attached file.
[BigDefinition.json](https://github.com/protocolbuffers/protobuf/files/14505963/BigDefinition.json)

And the results are:
main branch:
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3235/23H2/2023Update/SunValley3)
11th Gen Intel Core i9-11900K 3.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

Job=DefaultJob  Error=3.84 ns  StdDev=3.21 ns

| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     281.1 ns |   1.01 KB |
| Big    | 256,221.4 ns |  642.7 KB |

After:
| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     310.4 ns |     864 B |
| Big    | 259,694.1 ns |  536952 B |

In general I would say that it might be acceptable due to lower pressure over GC, but it is up to you to decide.
Or if you can check again what can be improved - I will gladly change code.

@amanda-tarafa FYI