@jskeet 2024-03-06
Given that the improvement in memory is small - and this *is* an ideal case for the GC, in terms of being able to collect everything from gen0 - I think it would probably be better to leave it as it is.

Thanks for trying it though - it's always hard to predict what performance effects will be without trying them. (And it does *sound* like a sensible change.)