@TrayanZapryanov 2024-03-06
Can I leave it open for a few days more?
I would like to try few tricks to see if possible to shave few milliseconds .