@jskeet 2024-03-06
Absolutely - I have no objections to that.