@TrayanZapryanov 2024-03-15
@jskeet I've found a nice nugets that can be used to verify your code for correct usage of structs and few other common mistakes.
[ErrorProne.NET.CoreAnalyzers](https://www.nuget.org/packages/ErrorProne.NET.CoreAnalyzers/0.6.1-beta.1)
[ErrorProne.NET.Structs](https://www.nuget.org/packages/ErrorProne.NET.Structs/0.6.1-beta.1)

This gave me some possibilities to optimize structs and I will try applying it's suggestions and to measure again why we have drop in performance. But also it found several more not related issues that might be interesting for you:

![image](https://github.com/protocolbuffers/protobuf/assets/12857389/8285d263-779d-4088-9244-4627fb7f5ddc)